assignment2

+++ SchoolManagementSystem/StaffDetails.cs
@@ -6,10 +6,15 @@ namespace SchoolManagementSystem
 {
     class StaffDetails
     {
+        #region Properties
         public string StaffName { get; set; }
         public DateTime DOB { get; set; }
-        public int MobileNumber { get; set; }
+        public decimal MobileNumber { get; set; }
         public string Address { get; set; }
         public int StaffID { get; set; }
+        #endregion
+
+        #region Methods
+
     }
 }
